* Import latest Nibble's changes on dietelf   - Adds support for exported symbols on non-stripped bins   - Code cleanup   - Shows and flags block sizes for each symbol * Warnatooon!!1   - Should fix any bugs * Moar advances in the waf port

+++ src/environ.c
@@ -132,7 +132,7 @@ void update_environment()
 
 	/* radis_update(); */
 
-	config.verbose = config_get("cfg.verbose");
+	config.verbose = (int)config_get("cfg.verbose");
 
 	ptr = getenv("COLUMNS");
 	if (ptr != NULL)
@@ -142,5 +142,5 @@ void update_environment()
 	if (config.width<10)config.width=10;
 	config_set_i("scr.width", config.width);
 
-	config.color = config_get("scr.color");
+	config.color = (int)config_get("scr.color");
 }
